requirements.txt
alembic==1.4.3
bcrypt==3.2.0
cffi==1.14.3
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT-Extended==3.24.1
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-Reuploaded==0.3.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.8.0
marshmallow-sqlalchemy==0.23.1
pkg-resources==0.0.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.19
Werkzeug==1.0.1
